changes per review comment
