+++ abods-api/template.yaml
@@ -132,7 +132,7 @@ Resources:
             - !Sub "{{resolve:ssm:/${ProjectName}/${Environment}/rds/dbname}}"
             - !Ref ProjectName
           DB_USER: !Sub "${ProjectName}_proxy_rw"
-          DD_API_KEY: !Sub "{{resolve:secretsmanager:${ProjectName}/${Environment}/datadog/apikey:SecretString}}"
+          DD_API_KEY_SECRET_ARN: !Sub "{{resolve:secretsmanager:${ProjectName}/${Environment}/datadog/apikey:SecretString}}"
           DD_SITE: datadoghq.eu
           M2M_API_SECRET_NAME: !Ref GraphQLAPITokenSecret
           DD_TAGS: app:graphql,project:abods
